refactor: move light client logic into separate files
